Migrate qos and mirror ct's to topology modular test framework
This includes updating the tests to use the vtysh communication library.
Doing so improves the reliability of the tests from ~50% to 94%. The 6% of
the time when they do fail is right away in the openswitch_setup.py: Timed
out while waiting for cur_cfg, which this repo has no control over.

Change-Id: Ie3df58880a94cd1aeaadfeee2c193de6c452e0b4
Signed-off-by: Miles Cote <miles.cote@hpe.com>